ml_api_project/requirements.txt
boto3==1.38.17
celery==5.5.2
dill==0.4.0
fastapi==0.115.12
pandas==2.2.3
pydantic==2.11.4
python-dotenv==1.1.0
scikit_learn==1.6.1
uvicorn==0.34.2
pydantic-settings==2.9.1
redis==6.1.0
